+++ contracts/smart-account-interfaces/src/lib.rs
@@ -1,7 +1,12 @@
 #![no_std]
 
+pub mod account;
 pub mod plugin;
 pub mod policy;
 
+pub use account::{
+    Ed25519Signer, ExternalPolicy, Secp256r1Signer, Signer, SignerKey, SignerPolicy, SignerRole,
+    SmartAccountClient,
+};
 pub use plugin::{SmartAccountPlugin, SmartAccountPluginClient};
 pub use policy::{SmartAccountPolicy, SmartAccountPolicyClient};
