+++ contracts/smart-account/src/auth/core/authorizer.rs
@@ -1,7 +1,7 @@
 /// Authorization service that verifies proofs and enforces role/policy checks.
 use crate::auth::permissions::{AuthorizationCheck, SignerRole};
 use crate::auth::proof::SignatureProofs;
-use crate::auth::signer::{Signer, SignerKey};
+use crate::auth::signer::{Signer, SignerKey, SignerExt};
 use crate::auth::signers::SignatureVerifier as _;
 use crate::config::{PLUGINS_KEY, TOPIC_PLUGIN, VERB_AUTH_FAILED};
 use crate::error::Error;
