+++ contracts/smart-account/src/auth/policy/mod.rs
@@ -1,8 +1,6 @@
 pub mod external;
 pub mod interface;
-pub mod time_based;
 
 pub use external::ExternalPolicy;
 pub use interface::SmartAccountPolicy;
 pub use interface::SmartAccountPolicyClient;
-pub use time_based::TimeBasedPolicy;
