+++ contracts/smart-account/src/auth/signer.rs
@@ -1,24 +1,10 @@
 use crate::auth::permissions::{AuthorizationCheck, SignerRole};
 use crate::auth::proof::SignerProof;
 use crate::auth::signers::SignatureVerifier;
-use crate::auth::signers::{Ed25519Signer, Secp256r1Signer};
 use crate::error::Error;
+pub use smart_account_interfaces::{Ed25519Signer, Secp256r1Signer, Signer, SignerKey};
 use soroban_sdk::Vec;
-use soroban_sdk::{auth::Context, contracttype, Bytes, BytesN, Env};
-
-#[contracttype]
-#[derive(Clone, Debug, PartialEq)]
-pub enum SignerKey {
-    Ed25519(BytesN<32>),
-    Secp256r1(Bytes),
-}
-
-#[contracttype]
-#[derive(Clone, Debug, PartialEq)]
-pub enum Signer {
-    Ed25519(Ed25519Signer, SignerRole),
-    Secp256r1(Secp256r1Signer, SignerRole),
-}
+use soroban_sdk::{auth::Context, BytesN, Env};
 
 impl SignatureVerifier for Signer {
     fn verify(&self, env: &Env, payload: &BytesN<32>, proof: &SignerProof) -> Result<(), Error> {
@@ -35,20 +21,22 @@ impl AuthorizationCheck for Signer {
     }
 }
 
-impl From<Signer> for SignerKey {
-    fn from(signer: Signer) -> Self {
-        match signer {
-            Signer::Ed25519(signer, _) => signer.into(),
-            Signer::Secp256r1(signer, _) => signer.into(),
-        }
-    }
+pub trait SignerExt {
+    fn role(&self) -> SignerRole;
 }
 
-impl Signer {
-    pub fn role(&self) -> SignerRole {
+impl SignerExt for Signer {
+    fn role(&self) -> SignerRole {
         match self {
             Signer::Ed25519(_, role) => role.clone(),
             Signer::Secp256r1(_, role) => role.clone(),
         }
+    }
+}
+
+pub fn signer_key_of(signer: &Signer) -> SignerKey {
+    match signer {
+        Signer::Ed25519(s, _) => SignerKey::Ed25519(s.public_key.clone()),
+        Signer::Secp256r1(s, _) => SignerKey::Secp256r1(s.key_id.clone()),
     }
 }
