+++ contracts/smart-account/src/auth/signers/mod.rs
@@ -1,8 +1,6 @@
 mod ed25519;
 pub mod secp256r1;
-
-pub use ed25519::Ed25519Signer;
-pub use secp256r1::Secp256r1Signer;
+pub use smart_account_interfaces::account::{Ed25519Signer, Secp256r1Signer};
 
 use crate::auth::proof::SignerProof;
 use crate::error::Error;
