+++ contracts/smart-account/src/auth/signers/secp256r1.rs
@@ -1,22 +1,11 @@
-use crate::auth::proof::{Secp256r1Signature, SignerProof};
-use crate::auth::signer::SignerKey;
 use crate::auth::signers::SignatureVerifier;
 use crate::error::Error;
 use base64ct::{Base64UrlUnpadded, Encoding};
-use soroban_sdk::{contracttype, Bytes, BytesN, Env};
+use smart_account_interfaces::Secp256r1Signer;
+use crate::auth::proof::{Secp256r1Signature, SignerProof};
+use soroban_sdk::{BytesN, Env};
 
-#[contracttype]
-#[derive(Clone, Debug, PartialEq)]
-pub struct Secp256r1Signer {
-    pub key_id: Bytes,
-    pub public_key: BytesN<65>,
-}
-
-impl Secp256r1Signer {
-    pub fn new(key_id: Bytes, public_key: BytesN<65>) -> Self {
-        Self { key_id, public_key }
-    }
-}
+// Secp256r1Signer type is now imported from smart_account_interfaces
 
 #[derive(serde::Deserialize)]
 struct ClientDataJson<'a> {
@@ -74,8 +63,4 @@ impl SignatureVerifier for Secp256r1Signer {
     }
 }
 
-impl From<Secp256r1Signer> for SignerKey {
-    fn from(signer: Secp256r1Signer) -> Self {
-        SignerKey::Secp256r1(signer.key_id.clone())
-    }
-}
+// SignerKey construction is handled via helper in signer module
