+++ contracts/smart-account/src/tests/policy_test.rs
@@ -9,7 +9,7 @@ use soroban_sdk::{
 
 use crate::account::SmartAccount;
 use crate::auth::permissions::{SignerPolicy, SignerRole};
-use crate::auth::policy::{ExternalPolicy, TimeBasedPolicy};
+use crate::auth::policy::ExternalPolicy;
 use crate::auth::signer::{Signer, SignerKey};
 use crate::error::Error;
 use crate::tests::test_utils::{setup, Ed25519TestSigner};
@@ -63,68 +63,6 @@ fn ensure_policy_event_is_not_emmited(env: &Env, policy_id: Address, event_name:
                 .unwrap_or(false)
         })
     }))
-}
-
-//
-// Time-based policy
-//
-#[test]
-fn test_deploy_with_time_based_policy() {
-    let env = setup();
-    let policy = SignerPolicy::TimeWindowPolicy(TimeBasedPolicy {
-        not_before: env.ledger().timestamp(),
-        not_after: env.ledger().timestamp() + 1000,
-    });
-    let admin_signer = Ed25519TestSigner::generate(SignerRole::Admin).into_signer(&env);
-    let test_signer =
-        Ed25519TestSigner::generate(SignerRole::Standard(vec![&env, policy])).into_signer(&env);
-    env.register(
-        SmartAccount,
-        (
-            vec![&env, admin_signer, test_signer],
-            Vec::<Address>::new(&env),
-        ),
-    );
-}
-
-#[test]
-#[should_panic]
-fn test_deploy_with_time_based_policy_wrong_time_range() {
-    let env = setup();
-    let policy = SignerPolicy::TimeWindowPolicy(TimeBasedPolicy {
-        not_before: env.ledger().timestamp() + 1000,
-        not_after: env.ledger().timestamp() + 999,
-    });
-    let admin_signer = Ed25519TestSigner::generate(SignerRole::Admin).into_signer(&env);
-    let test_signer =
-        Ed25519TestSigner::generate(SignerRole::Standard(vec![&env, policy])).into_signer(&env);
-    env.register(
-        SmartAccount,
-        (
-            vec![&env, admin_signer, test_signer],
-            Vec::<Address>::new(&env),
-        ),
-    );
-}
-
-#[test]
-#[should_panic]
-fn test_deploy_with_time_based_policy_wrong_not_after() {
-    let env = setup();
-    let policy = SignerPolicy::TimeWindowPolicy(TimeBasedPolicy {
-        not_before: env.ledger().timestamp() + 1000,
-        not_after: env.ledger().timestamp() + 999,
-    });
-    let admin_signer = Ed25519TestSigner::generate(SignerRole::Admin).into_signer(&env);
-    let test_signer =
-        Ed25519TestSigner::generate(SignerRole::Standard(vec![&env, policy])).into_signer(&env);
-    env.register(
-        SmartAccount,
-        (
-            vec![&env, admin_signer, test_signer],
-            Vec::<Address>::new(&env),
-        ),
-    );
 }
 
 #[test]
